requirements.txt
streamlit
pymupdf
faiss-cpu
python-dotenv
google-generativeai>=0.8.2
isort
ruff